The density function for parasite densities in a simple malaria infection of age alpha

The density function for parasite densities in a simple malaria infection of age alpha

Usage

d_alpha2density(
  xi,
  alpha,
  W = 0,
  a = 0,
  par_RBC = par_lRBC_static(),
  par_Fmu = par_Fmu_base(),
  par_Omega = par_Omega_beta()
)

Arguments

xi

a vector of quantiles

alpha

the age of a parasite infection

W

the immune tracking variables

a

host cohort age

par_RBC

parameters to compute log10RBC

par_Fmu

parameters to compute Fmu

par_Omega

parameters to compute d_Omega

Value

a numeric vector of length(x)