The random generation function for the beta distribution, an alternative parameterization Title

The random generation function for the beta distribution, an alternative parameterization Title

Usage

rbeta1(n, mu, pSig = par_sigma_abc())

Arguments

n

number of observations

mu

the mean value for the distribution (0 <= mu <= 1)

pSig

parameters to dispatch the S3 function sigma_mu

Value

a numeric vector of length n